airbnbscrap/requirements.txt
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==3.2
django-cron==0.6.0
djangorestframework==3.14.0
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
lxml==4.9.2
numpy==1.24.0
outcome==1.2.0
pandas==1.5.2
psycopg2==2.9.5
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
wsproto==1.2.0
